✨ Use public always-online STUN server list
STUN servers may come and go all the time, and Hyphanet should try to remain as
inconspicuous as possible. Using a publicly available and constantly updated
list of available STUN servers should reduce the possibility of detection that
Hyphanet with the JSTUN plugin is being run. As the list that has currently
been implemented seems to be used by many different people, neither the request
to Github for the list, nor requests to STUN servers on that list should make
Hyphanet stand out in any way.